+++ client/src/commands/restartFabricRuntime.ts
@@ -17,6 +17,7 @@ import { VSCodeBlockchainOutputAdapter } from '../logging/VSCodeBlockchainOutput
 import { FabricRuntime } from '../fabric/FabricRuntime';
 import { FabricRuntimeManager } from '../fabric/FabricRuntimeManager';
 import { LogType } from '../logging/OutputAdapter';
+import { FabricRuntimeUtil } from '../fabric/FabricRuntimeUtil';
 
 export async function restartFabricRuntime(): Promise<void> {
     const outputAdapter: VSCodeBlockchainOutputAdapter = VSCodeBlockchainOutputAdapter.instance();
@@ -28,12 +29,16 @@ export async function restartFabricRuntime(): Promise<void> {
         title: 'IBM Blockchain Platform Extension',
         cancellable: false
     }, async (progress: vscode.Progress<{ message: string }>) => {
-        progress.report({ message: `Restarting Fabric runtime ${runtime.getName()}` });
+        let fabricRuntimeName: string = runtime.getName();
+        if (fabricRuntimeName === FabricRuntimeUtil.LOCAL_FABRIC) {
+            fabricRuntimeName = FabricRuntimeUtil.LOCAL_FABRIC_DISPLAY_NAME;
+        }
+        progress.report({ message: `Restarting Fabric runtime ${fabricRuntimeName}` });
 
         try {
             await runtime.restart(outputAdapter);
         } catch (error) {
-            outputAdapter.log(LogType.ERROR, `Failed to restart local_fabric: ${error.message}`, `Failed to restart local_fabric: ${error.toString()}`);
+            outputAdapter.log(LogType.ERROR, `Failed to restart Local Fabric: ${error.message}`, `Failed to restart Local Fabric: ${error.toString()}`);
         }
     });
 }