+++ client/src/commands/toggleFabricRuntimeDevMode.ts
@@ -19,6 +19,7 @@ import { FabricConnectionManager } from '../fabric/FabricConnectionManager';
 import { FabricRuntimeManager } from '../fabric/FabricRuntimeManager';
 import { LogType } from '../logging/OutputAdapter';
 import { ExtensionCommands } from '../../ExtensionCommands';
+import { FabricRuntimeUtil } from '../fabric/FabricRuntimeUtil';
 
 export async function toggleFabricRuntimeDevMode(): Promise<void> {
     const outputAdapter: VSCodeBlockchainOutputAdapter = VSCodeBlockchainOutputAdapter.instance();
@@ -41,11 +42,15 @@ export async function toggleFabricRuntimeDevMode(): Promise<void> {
             title: 'IBM Blockchain Platform Extension',
             cancellable: false
         }, async (progress: vscode.Progress<{ message: string }>) => {
-            progress.report({ message: `Restarting Fabric runtime ${runtime.getName()}` });
+            let fabricRuntimeName: string = runtime.getName();
+            if (fabricRuntimeName === FabricRuntimeUtil.LOCAL_FABRIC) {
+                fabricRuntimeName = FabricRuntimeUtil.LOCAL_FABRIC_DISPLAY_NAME;
+            }
+            progress.report({ message: `Restarting Fabric runtime ${fabricRuntimeName}` });
             try {
                 await runtime.restart(outputAdapter);
             } catch (error) {
-                outputAdapter.log(LogType.ERROR, `Failed to restart local_fabric: ${error.message}`, `Failed to restart local_fabric: ${error.toString()}`);
+                outputAdapter.log(LogType.ERROR, `Failed to restart Local Fabric: ${error.message}`, `Failed to restart Local Fabric: ${error.toString()}`);
                 return;
             }
         });